requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
jinja2==3.1.4
python-multipart==0.0.20
